Moved serial, tcp & modbus classes to separate plugin.
