+++ lib/daq/QDaqInterface.h
@@ -4,12 +4,8 @@
 #include "QDaqObject.h"
 
 #include <QVector>
-#include <QHostAddress>
-#include <QSerialPort>
 
 class QDaqDevice;
-
-class tcp_socket;
 
 /**
  * @brief Base class of objects representing digital communications interfaces.
@@ -157,223 +153,10 @@ public slots:
 };
 
 
-/**
- * @brief A class for accesing a TCP/IP interface.
- *
- * @ingroup Daq
- * @ingroup ScriptAPI
- *
- * QDaqTcpip performs communications with another host
- * using a tcp socket connection
- * to the specified host address and port. The port here
- * means the tcp/ip port and is not to be confused with
- * the ports of QDaqInterface.
- *
- * QDaqTcpip can be used in order to communicate with a single
- * device. For multiple devices connected on the ethernet bus, each with its
- * own host ip address,
- * a separate QDaqTcpip must be used per device.
- *
- * If the same device with single ip address offers several ports
- * for communication, then a separate QDaqTcpip must be used per
- * device port.
- *
- */
-class QDAQ_EXPORT QDaqTcpip : public QDaqInterface
-{
-	Q_OBJECT
-
-    /// Tcp/ip port.
-	Q_PROPERTY(uint port READ port WRITE setPort)
-    /// Host IP address as string, eg '192.168.1.6'
-	Q_PROPERTY(QString host READ host WRITE setHost)
-
-protected:
-    // properties
-	uint port_;
-	QHostAddress host_;
-    // my socket
-    tcp_socket* socket_;
-public:
-
-    /**
-     * @brief QDaqTcpip constructor.
-     *
-     * It initializes the object with the interface is in the closed state.
-     *
-     * @param name QDaqObject name
-     * @param ahost Host address
-     * @param portno Port number where the host listens
-     */
-    Q_INVOKABLE explicit QDaqTcpip(const QString& name, const QString& ahost = QString(), uint portno = 0);
-    virtual ~QDaqTcpip();
-
-	// getters
-	uint port() const { return port_; }
-	QString host() const { return host_.toString(); }
-
-	// setters
-	void setPort(uint p);
-	void setHost(const QString& h);
 
 
-	virtual bool isValidPort(uint i) { return i==0; }
 
-	// io
-	virtual int read(uint port, char* buff, int len, int eos = 0);
-	virtual int write(uint port, const char* buff, int len, int eos = 0);
 
-protected:
-	virtual bool open_();
-	virtual void close_();
-};
-/**
- * @brief A class for accesing a serial bus.
- *
- * @ingroup Daq
- * @ingroup ScriptAPI
- *
- * QDaqSerial performs communications with devices connected
- * to a serial port. It is essentially a wrapper for the QSerialPort class.
- * Communications are performed in blocking mode, i.e., QDaq will wait for
- * a message to be read/written to the serial port before it continues.
- * This is because interfaces will be generally used in a QDaqLoop that runs
- * in a separate thread and in most cases the information that must be transmitted/obtained
- * from the serial port is essential for the loop to continue.
- *
- *
- */
-class QDAQ_EXPORT QDaqSerial : public QDaqInterface
-{
-    Q_OBJECT
-
-    /// Name of the serial port, eg, 'COM1' in windows, '/dev/ttyS0' in linux.
-    Q_PROPERTY(QString portName READ portName WRITE setPortName)
-    /// Communication bit rate in baud
-    Q_PROPERTY(QSerialPort::BaudRate baud READ baud WRITE setBaud)
-    /// The parity to be used
-    Q_PROPERTY(QSerialPort::Parity parity READ parity WRITE setParity)
-    /// Number of data bits
-    Q_PROPERTY(QSerialPort::DataBits databits READ databits WRITE setDatabits)
-    /// Number of stop bits
-    Q_PROPERTY(QSerialPort::StopBits stopbits READ stopbits WRITE setStopbits)
-    /// Type of flow control
-    Q_PROPERTY(QSerialPort::FlowControl handshake READ handshake WRITE setHandshake)
-
-protected:
-    QSerialPort* port_;
-
-public:
-    Q_INVOKABLE
-    /**
-     * @brief QDaqSerial constructor.
-     * @param name QDaqObject name
-     * @param portName OS name for the serial port, eg, 'COM1' in windows, '/dev/ttyS0' in linux
-     */
-    explicit QDaqSerial(const QString& name, const QString& portName = QString());
-
-    // getters
-    QString portName() const { return port_->portName(); }
-    QSerialPort::BaudRate baud() const { return (QSerialPort::BaudRate)port_->baudRate(); }
-    QSerialPort::Parity parity() const { return port_->parity(); }
-    QSerialPort::DataBits databits() const { return port_->dataBits(); }
-    QSerialPort::StopBits stopbits() const { return port_->stopBits(); }
-    QSerialPort::FlowControl handshake() const { return port_->flowControl(); }
-
-    // setters
-    void setPortName(const QString& aname);
-    void setBaud(QSerialPort::BaudRate v);
-    void setParity(QSerialPort::Parity v);
-    void setDatabits(QSerialPort::DataBits v);
-    void setStopbits(QSerialPort::StopBits v);
-    void setHandshake(QSerialPort::FlowControl v);
-
-    // io
-    virtual int read(uint port, char* buff, int len, int eos = 0);
-    virtual int write(uint port, const char* buff, int len, int eos = 0);
-
-protected:
-    virtual bool open_();
-    virtual void close_();
-    virtual void clear_();
-};
-/**
- * @brief A class for accesing a MODBUS/TCP connection.
- *
- * @ingroup Daq
- * @ingroup ScriptAPI
- *
- * Used to access MODBUS devices with the TCP/IP protocol.
- * A separate QDaqModbusTcp object must be used per device.
- *
- * Port number is typically 502.
- *
- * QDaqModbusTcp can read/write only MODBUS registers, i.e. 16 bit words.
- *
- * Implementation of MODBUS in QDaq is based on libmodbus.
- *
- */
-class QDAQ_EXPORT QDaqModbusTcp : public QDaqTcpip
-{
-    Q_OBJECT
-
-protected:
-    void* ctx_;
-
-public:     
-    /**
-     * @brief QDaqModbusTcp constructor
-     * @param name The QDaqObject name
-     * @param ahost Host address as string, eg, '192.168.1.2'
-     * @param portno port number, typically 502
-     */
-    Q_INVOKABLE explicit QDaqModbusTcp(const QString& name, const QString& ahost = QString(), uint portno = 502);
-    virtual ~QDaqModbusTcp();
-
-    // io
-    virtual int read(uint port, char* buff, int len, int eos = 0);
-    virtual int write(uint port, const char* buff, int len, int eos = 0);
-
-protected:
-    virtual bool open_();
-    virtual void close_();
-};
-/**
- * @brief A class for accesing a MODBUS/RTU connection.
- *
- * @ingroup Daq
- * @ingroup ScriptAPI
- *
- * Used to access MODBUS devices connected to a serial port.
- * A separate QDaqModbusRtu object must be used per device.
- *
- * Implementation of MODBUS in QDaq is based on libmodbus.
- *
- */
-class QDAQ_EXPORT QDaqModbusRtu : public QDaqSerial
-{
-    Q_OBJECT
-
-protected:
-    void* ctx_;
-
-public:     
-    /**
-     * @brief QDaqModbusRtu constructor
-     * @param name QDaqObject name
-     * @param portName OS name for the serial port, eg, 'COM1' in windows, '/dev/ttyS0' in linux.
-     */
-    Q_INVOKABLE explicit QDaqModbusRtu(const QString& name, const QString& portName = QString());
-    virtual ~QDaqModbusRtu();
-
-    // io
-    virtual int read(uint port, char* buff, int len, int eos = 0);
-    virtual int write(uint port, const char* buff, int len, int eos = 0);
-
-protected:
-    virtual bool open_();
-    virtual void close_();
-};
