+++ plugins/interfaces/qdaqinterfaces.json
@@ -1,0 +1,1 @@
+{}
