+++ plugins/plugins.pro
@@ -5,5 +5,6 @@ unix: SUBDIRS +=  linux-gpib
 win32: SUBDIRS +=  ni-gpib
 
 SUBDIRS += \
-    filters
+    filters \
+    interfaces
